Реализует оффлайн режим

+++ src/api/provider.js
@@ -1,21 +1,88 @@
+import Film from "../models/film";
+
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
+const getSyncedFilms = (items) => {
+  return items.filter(({success}) => success)
+    .map(({payload}) => payload.movie);
+};
+
+const createStoreStructure = (items) => {
+  return items.reduce((acc, current) => {
+    return Object.assign({}, acc, {
+      [current.id]: current,
+    });
+  }, {});
+};
+
 export default class Provider {
-  constructor(api) {
+  constructor(api, store) {
     this._api = api;
+    this._store = store;
   }
 
   getFilms() {
-    return this._api.getFilms();
+    if (isOnline()) {
+      return this._api.getFilms()
+        .then((films) => {
+          const items = createStoreStructure(films.map((film) => film.toRAW()));
+
+          this._store.setItems(items);
+
+          return films;
+        });
+    }
+
+    const storeFilms = Object.values(this._store.getItems());
+    return Promise.resolve(Film.parseMultiple(storeFilms));
   }
 
   updateFilm(id, film) {
-    return this._api.updateFilm(id, film);
+    if (isOnline()) {
+      return this._api.updateFilm(id, film)
+        .then((newFilm) => {
+          this._store.setItem(newFilm.id, newFilm.toRAW());
+
+          return newFilm;
+        });
+    }
+    const localFilm = Film.clone(Object.assign(film, {id}));
+    this._store.setItem(id, localFilm.toRAW());
+    return Promise.resolve(localFilm);
   }
 
   createComment(id, comment) {
-    return this._api.createComment(id, comment);
+    if (isOnline()) {
+      return this._api.createComment(id, comment);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deleteComment(id) {
-    return this._api.deleteComment(id);
+    if (isOnline()) {
+      return this._api.deleteComment(id);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
+  }
+
+  sync() {
+    if (isOnline()) {
+      const storeFilms = Object.values(this._store.getItems());
+
+      return this._api.sync(storeFilms)
+        .then((response) => {
+          const updatedFilms = getSyncedFilms(response.updated);
+
+          const items = createStoreStructure([...updatedFilms]);
+
+          this._store.setItems(items);
+        });
+    }
+
+    return Promise.reject(new Error(`Sync data failed`));
   }
 }
